temperature fix1

+++ if.py
@@ -1,4 +1,4 @@
 celsius = input('攝氏溫度：')
-celsius = int(celsius)
+celsius = float(celsius)
 Fahrenheit = celsius * (9 / 5) + 32
 print('華氏溫度:', Fahrenheit)
